requirements.txt
metaphor-python
unidecode
beautifulsoup4
requests
python-dotenv
nltk